Breaking change: updating the proxy assumptions
If the cors proxy was entered before, it was assumed that there should
be a slash between the proxy URL and the URLs to be fetched. This is
changing and it makes the assumption that the cors proxy provided is
entirely correct, and the URL should be appended directly to that.